rewizje-zmiana-typy-danych-klucza

+++ database/migrations/2023_05_31_084504_create_rewizje_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->boolean('Czy_uszkodzona');
             $table->string('Opis');
             $table->decimal('Koszt_uszkodzen');
-            $table->unsignedBigInteger('hulajnoga_id');
+            $table->unsignedInteger('hulajnoga_id');
             $table->foreign('hulajnoga_id')->references('id')->on('hulajnogi');
             $table->timestamps();
         });
